milestone03: correct id on PokemonFacade

+++ pkmnleague-api/src/main/java/cz/fi/muni/pa165/seminar/pkmnleague/facade/PokemonFacade.java
@@ -44,7 +44,7 @@ public interface PokemonFacade {
         * @param id
         * @return PokemonDTO
         */
-        public PokemonDTO getPokemonWithId(long id);
+        public PokemonDTO getPokemonWithId(int id);
         
         /**
         * Gets existing Pokemons by Trainer ID
